mobile responsiveness, redesign folder structure
